pyproject.toml
[project]
name = "repo2prompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataclasses>=0.8",
    "pathlib>=1.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.8.1",
]
